Added newlines to the end of source files that didn't have them. Fixed a 'cast to pointer from integer of different size' warning in TreeStorageNode.cpp

+++ src/TreeStorageNode.cpp
@@ -76,10 +76,10 @@ void* TreeStorageNode::getNextAttribute(void* pUserData,std::string& name,std::v
 }
 
 void* TreeStorageNode::getNextNode(void* pUserData,ITreeStorageReader*& obj){
-	unsigned int i=(intptr_t)pUserData;
+	intptr_t i=(intptr_t)pUserData;
 	
 	//Check if the pointer is in range of the subNodes vector.
-	if(i<subNodes.size()){
+	if(i<(int)subNodes.size()){
 		obj=subNodes[i];
 		return (void*)(i+1);
 	}else{
